separating the structure viewer
