+++ src/components/Structure/ViewerAndEntries/ProtVistaForStructures.js
